date of workhour model add. andere date attributen weg.

+++ db/migrate/20131116133243_create_users.rb
@@ -2,8 +2,8 @@ class CreateUsers < ActiveRecord::Migration
   def change
     create_table :users do |t|
       t.string :name
-      t.double :brutoUurLoon
-      t.double :nettoUurLoon
+      t.decimal :brutoUurLoon
+      t.decimal :nettoUurLoon
 
       t.timestamps
     end
